make wlr_output_layout_output private
