+++ types/wlr_output_layout.c
@@ -7,21 +7,10 @@
 #include <wlr/types/wlr_output.h>
 #include <wlr/util/log.h>
 #include "util/signal.h"
+#include "types/wlr_output_layout.h"
 
 struct wlr_output_layout_state {
 	struct wlr_box _box; // should never be read directly, use the getter
-};
-
-struct wlr_output_layout_output_state {
-	struct wlr_output_layout *layout;
-	struct wlr_output_layout_output *l_output;
-
-	bool auto_configured;
-
-	struct wl_listener mode;
-	struct wl_listener scale;
-	struct wl_listener transform;
-	struct wl_listener output_destroy;
 };
 
 struct wlr_output_layout *wlr_output_layout_create(void) {
